Seperated modal logic into its own hook. everything still needs testing
